# swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger REST Article
consumes:
  - "application/json"
produces:
  - "application/json"

basePath: "/json_crud"

# Paths supported by the server application
paths:
  /addSubject:
      post:
          operationId: question.add_subject
          tags:
            - question
          summary: Create a question and add it to the question list
          description: Create a new question in the question list
          parameters:
            - name: question
              in: body
              description: question to create
              required: True
              schema:
                type: object
                properties:
                  subject_name:
                    type: string
                    description: Name of the Subject
          responses:
            201:
              description: Successfully created  subject in dictonary
#route add_question
  /add_question:
        post:
          operationId: question.add_question
          tags:
            - question
          summary: Create a question and add it to the question list
          description: Create a new question in the question list
          parameters:
            - name: question
              in: body
              description: question to create
              required: True
              schema:
                type: object
                properties:
                  subject_name:
                    type: string
                    description: Name of the Subject
                  question_number:
                    type: string
                    description: Number of the Question
                  question_discription:
                    type: string
                    description: Discription of the question
          responses:
            201:
              description: Successfully created  subject in dictonary
#route update_question
  /update_question:
      post:
          operationId: question.update_question
          tags:
            - question
          summary: Create a question and add it to the question list
          description: Create a new question in the question list
          parameters:
            - name: question
              in: body
              description: question to create
              required: True
              schema:
                type: object
                properties:
                  subject_name:
                    type: string
                    description: Name of the Subject
                  question_number:
                    type: string
                    description: Number of the Question
                  updated_question_discription:
                    type: string
                    description: Discription of the updated_question_discription
          responses:
            201:
              description: Successfully created  subject in dictonary
  #route delete_question
  /delete_question:
      post:
          operationId: question.delete_question
          tags:
            - question
          summary: Create a question and add it to the question list
          description: Create a new question in the question list
          parameters:
            - name: question
              in: body
              description: question to create
              required: True
              schema:
                type: object
                properties:
                  subject_name:
                    type: string
                    description: Name of the Subject
                  question_number:
                    type: string
                    description: Number of the Question
          responses:
            201:
              description: Successfully created  subject in dictonary
   #route delete_subject
  /delete_subject:
      post:
          operationId: question.delete_subject
          tags:
            - question
          summary: Create a question and add it to the question list
          description: Create a new question in the question list
          parameters:
            - name: question
              in: body
              description: question to create
              required: True
              schema:
                type: object
                properties:
                  subject_name:
                    type: string
                    description: Name of the Subject
          responses:
            201:
              description: Successfully created  subject in dictonary
  #getSubjects
  /getSubjects:
    get:
      operationId: question.get_subjects
      tags:
        - "question"
      summary: "The question data structure supported by the server application"
      description: "Read the list of Subjects"
      responses:
        200:
          description: "Successful read list of subjects"
          schema:
            type: "array"
            items:
              properties:
                subject:
                  type: "string"
  #getQuestion
  /getSubjectQuestions:
    get:
      operationId: question.get_subjects_questions
      tags:
        - "question"
      summary: "The question data structure supported by the server application"
      description: "Read the list of question"
      responses:
        200:
          description: "Successful read list of question"
          schema:
            type: "array"
            items:
              properties:
                subject:
                  type: "string"
      